+++ simba/core/models/simba_model.py
@@ -7,7 +7,7 @@ from torch.utils.data import DataLoader
 from simba.analog_discovery.fc_layers_analog_discovery import (
     FcLayerAnalogDiscovery,
 )
-from simba.core.data.encoder_dataset_builder import prepare_encoder_dataset
+from simba.core.data.datasets.encoder_dataset_builder import prepare_encoder_dataset
 from simba.core.models.embedder_multitask import EmbedderMultitask
 from simba.core.models.encoder import Encoder
