+++ simba/workflows/inference.py
@@ -15,8 +15,8 @@ from torch.utils.data import DataLoader
 import simba.core.data.molecule_pairs
 import simba.core.data.spectrum
 from simba.core.chemistry.mces_loader.load_mces import LoadMCES
+from simba.core.data.datasets.multitask_dataset_builder import MultitaskDataBuilder
 from simba.core.data.molecule_pairs import MoleculePairsOpt
-from simba.core.data.multitask_dataset_builder import MultitaskDataLoader
 from simba.core.models.embedder_multitask import EmbedderMultitask
 from simba.core.training.train_utils import TrainUtils
 from simba.utils.logger_setup import logger
@@ -150,7 +150,7 @@ def prepare_inference_dataloaders(
 
     # Create dataloaders
     logger.info("Creating dataloaders...")
-    dataset_ed = MultitaskDataLoader.from_molecule_pairs_to_dataset(
+    dataset_ed = MultitaskDataBuilder.from_molecule_pairs_to_dataset(
         molecule_pairs_ed_uniform,
         max_num_peaks=int(cfg.model.transformer.context_length),
         use_adduct=cfg.model.features.use_adduct,
@@ -162,7 +162,7 @@ def prepare_inference_dataloaders(
         dataset_ed, batch_size=cfg.inference.batch_size, shuffle=False
     )
 
-    dataset_mces = MultitaskDataLoader.from_molecule_pairs_to_dataset(
+    dataset_mces = MultitaskDataBuilder.from_molecule_pairs_to_dataset(
         molecule_pairs_mces_uniform,
         max_num_peaks=int(cfg.model.transformer.context_length),
         use_adduct=cfg.model.features.use_adduct,
